Redesigned layout implemented, has bugs but looks OK, great progress
